requirements.txt
requests>=2.31.0
pytz>=2024.1
openai>=1.14.0
google-generativeai>=0.4.0